Refactor backend for D2C Men's Clothing, Stripe integration, and Docker setup

+++ src/Ecommerce.Application/DTOs/Cart/CartItemResponseDto.cs
@@ -4,9 +4,9 @@ namespace Ecommerce.Application.DTOs.Cart
     {
         public Guid ProductId { get; set; }
         public string ProductName { get; set; } = null!;
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public string Image { get; set; } = null!;
     }
 }
