+++ src/Ecommerce.Application/DTOs/Orders/OrderItemResponseDto.cs
@@ -6,8 +6,8 @@ namespace Ecommerce.Application.DTOs.Orders
         public Guid ProductId { get; set; }
         public string ProductName { get; set; } = null!;
         public string ImageUrl { get; set; } = null!;
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
-        public int TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
