+++ src/Ecommerce.Application/DTOs/Orders/RevenueResponseDto.cs
@@ -2,7 +2,7 @@ namespace Ecommerce.Application.DTOs.Orders
 {
     public class RevenueResponseDto
     {
-        public int TotalRevenue { get; set; }
+        public decimal TotalRevenue { get; set; }
         public int TotalItemsSold { get; set; }
     }
 }
